+++ packages/bluesky/example/example.dart
@@ -59,7 +59,7 @@ Future<void> main() async {
 Future<bsky.Session> get _session async {
   final session = await bsky.createSession(
     service: 'SERVICE_NAME', //! The default is `bsky.social`
-    handle: 'YOUR_HANDLE', //! Like `shinyakato.bsky.social`
+    identifier: 'YOUR_HANDLE_OR_EMAIL', //! Like `shinyakato.bsky.social`
     password: 'YOUR_PASSWORD',
   );