+++ packages/bluesky_cards/example/lib/main.dart
@@ -24,7 +24,7 @@ class BlueskyCards extends StatelessWidget {
     return Scaffold(
       body: FutureBuilder(
         future: bsky.createSession(
-          handle: 'YOUR_HANDLE',
+          handle: 'YOUR_HANDLE_OR_EMAIL',
           password: 'YOUR_PASSWORD',
         ),
         builder: (BuildContext context, AsyncSnapshot snapshot) {
@@ -37,7 +37,7 @@ class BlueskyCards extends StatelessWidget {
 
           return FutureBuilder(
             future: bluesky.feeds.findFeeds(
-              author: session.did,
+              actor: session.did,
               limit: 5,
             ),
             builder: (BuildContext context, AsyncSnapshot snapshot) {